Fix virtual layers suite internal handling
Since the issue that removed the QgsProject::instance set inside
the config cache, we still have an invalid virtual layer
because at the time the layer is created the project instance
is not yet set.

So, we are checking if there is any invalid virtual layer and
reload the project, this happens only once right after
the project is retrieved from the cache for the first time.