Do not show the subtitle `None` when the macrogroup does not have a name (yet) (#1152)
If one is adding a new macro group, the name is still None. In other cases
it can also have an empty value.

In such cases, we should not print anything in the template, otherwise
we see "None" as a subtitle.

(cherry picked from commit 655517ae8738d4658eb17e61cf53d94e56927923)